services/ingest-weather/requirements.txt
requests
psycopg2-binary
apscheduler
